Conditionally inject supportedCapabilities
